클래스 다이어그램 작성
