--- src/main/java/CarFactory.java
@@ -1,15 +1,0 @@
-import java.util.ArrayList;
-import java.util.List;
-
-public class CarFactory {
-    public CarFactory() {
-    }
-
-    public Cars make(int howManyCars) {
-        List<Car> cars = new ArrayList<>();
-        for (int i = 0 ; i < howManyCars; i++) {
-            cars.add(Car.create());
-        }
-        return new Cars(cars);
-    }
-}
