--- src/main/java/racing/Referee.java
@@ -1,7 +1,0 @@
-package racing;
-
-public interface Referee {
-    int STANDARD_NUMBER_TO_JUDGE_TO_MOVE = 4;
-
-    boolean judgeToMove();
-}
