+++ src/main/java/step3/CarFactory.java
@@ -1,0 +1,17 @@
+package step3;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class CarFactory {
+    public CarFactory() {
+    }
+
+    public Cars make(int howManyCars) {
+        List<Car> cars = new ArrayList<>();
+        for (int i = 0 ; i < howManyCars; i++) {
+            cars.add(Car.create());
+        }
+        return new Cars(cars);
+    }
+}
