+++ src/main/java/step4/racing/Referee.java
@@ -1,0 +1,7 @@
+package step4.racing;
+
+public interface Referee {
+    int STANDARD_NUMBER_TO_JUDGE_TO_MOVE = 4;
+
+    boolean judgeToMove();
+}
